monitoring-web/requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.0
msal==1.26.0
flask-session==0.5.0
pyodbc==5.1.0
